feat: migrate to @origin-cards/api from npm, remove packages/ (#phase6b)
- Installs @origin-cards/api@0.7.0 from npm registry
- Updates @origin/api → @origin-cards/api in src/ (pluginBus.ts, workspaceStore.ts)
- Moves OriginChannelMap, PluginBus, PluginContext, PluginManifest, PluginLifecycleEvent
  into src/types/plugin.ts (npm package lacks these richer types)
- Removes @origin/api, @origin/hello, @origin/notepad path aliases from tsconfig.json
- Removes @origin/api vite alias from vite.config.ts
- Removes workspaces field from package.json
- Deletes packages/ directory (all packages now live in origin-plugin-monorepo)

Typecheck: identical pre-existing errors only (App.tsx, settings.bak, workspaceStore filterKeysStrategy).
