--- packages/api/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@origin-cards/api",
-  "version": "0.1.0",
-  "description": "Type contract for Origin plugins — PluginManifest, PluginContext, PluginComponent, PluginModule",
-  "type": "module",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsup src/index.ts --format esm --dts --clean --external react",
-    "docs": "typedoc",
-    "prepublishOnly": "npm run build"
-  },
-  "peerDependencies": {
-    "react": ">=18"
-  },
-  "devDependencies": {
-    "@types/react": "^19.0.0",
-    "tsup": "^8.0.0",
-    "typedoc": "^0.27.9",
-    "typescript": "~5.7.2"
-  }
-}
