--- packages/api/src/index.ts
@@ -1,11 +1,0 @@
-export type {
-  OriginChannelMap,
-  PluginBus,
-  PluginManifest,
-  PluginContext,
-  PluginLifecycleEvent,
-  PluginComponent,
-  PluginModule,
-} from "./plugin";
-
-export { useBusChannel } from "./useBusChannel";
