--- packages/api/typedoc.json
@@ -1,8 +1,0 @@
-{
-  "entryPoints": ["src/index.ts"],
-  "tsconfig": "tsconfig.json",
-  "json": "dist/api-docs.json",
-  "readme": "none",
-  "excludeInternal": true,
-  "skipErrorChecking": true
-}
