--- packages/browser/src/manifest.json
@@ -1,7 +1,0 @@
-{
-  "id": "com.origin.browser",
-  "name": "Browser",
-  "version": "0.1.0",
-  "description": "View any URL in a panel — great for localhost dev servers",
-  "icon": "🌐"
-}
