--- packages/browser/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.browser",
-  name: "Browser",
-  version: "0.1.0",
-  description: "View any URL in a panel — great for localhost dev servers",
-  icon: "🌐",
-};
