--- packages/filetree/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@origin/filetree",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite build --watch"
-  },
-  "dependencies": {
-    "@origin/api": "*"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.4.1",
-    "vite": "^7.0.0"
-  }
-}
