--- packages/filetree/src/manifest.json
@@ -1,7 +1,0 @@
-{
-  "id": "com.origin.filetree",
-  "name": "File Tree",
-  "version": "0.1.0",
-  "description": "Browse your local filesystem",
-  "icon": "📁"
-}
