--- packages/filetree/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.filetree",
-  name: "File Tree",
-  version: "0.1.0",
-  description: "Browse your local filesystem",
-  icon: "📁",
-};
