--- packages/github/src/manifest.json
@@ -1,7 +1,0 @@
-{
-  "id": "com.origin.github",
-  "name": "GitHub PRs",
-  "version": "0.1.0",
-  "description": "Track open pull requests for any GitHub repository",
-  "icon": "🐙"
-}
