--- packages/github/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.github",
-  name: "GitHub PRs",
-  version: "0.1.0",
-  description: "Track open pull requests for any GitHub repository",
-  icon: "🐙",
-};
