--- packages/hello/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@origin/hello",
-  "version": "0.1.0",
-  "private": true,
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "dependencies": {
-    "@origin/api": "*"
-  }
-}
