--- packages/hello/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.hello",
-  name: "Hello",
-  version: "0.1.0",
-  description: "Reference plugin — minimal hello world for plugin authors.",
-  icon: "👋",
-};
