--- packages/monaco/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@origin/monaco",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite build --watch"
-  },
-  "dependencies": {
-    "@origin/api": "*",
-    "@monaco-editor/react": "^4.7.0"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.4.1",
-    "vite": "^7.0.0"
-  }
-}
