--- packages/monaco/src/manifest.json
@@ -1,7 +1,0 @@
-{
-  "id": "com.origin.monaco",
-  "name": "Monaco Editor",
-  "version": "0.1.0",
-  "description": "VS Code editor — open and edit any file",
-  "icon": "✏️"
-}
