--- packages/monaco/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.monaco",
-  name: "Monaco Editor",
-  version: "0.1.0",
-  description: "VS Code editor — open and edit any file",
-  icon: "✏️",
-};
