--- packages/notepad/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@origin/notepad",
-  "version": "0.1.0",
-  "private": true,
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "dependencies": {
-    "@origin/api": "*",
-    "@tauri-apps/plugin-fs": "^2"
-  }
-}
