--- packages/notepad/src/manifest.ts
@@ -1,9 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.notepad",
-  name: "Notepad",
-  version: "0.1.0",
-  description: "A simple markdown notepad — notes are saved per panel",
-  icon: "📝",
-};
