--- packages/sdk/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@origin/sdk",
-  "version": "0.1.0",
-  "private": true,
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "dependencies": {}
-}
