--- packages/sdk/src/index.ts
@@ -1,6 +1,0 @@
-export { usePluginContext } from "./context";
-export type { IframePluginContextWithConfig } from "./context";
-export { useBusChannel } from "./bus";
-export { invoke } from "./invoke";
-export { onEvent, useOriginEvent } from "./events";
-export type { IframePluginContext } from "@/lib/iframeProtocol";
