--- packages/template/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@origin/template",
-  "version": "0.1.0",
-  "private": true,
-  "deprecated": "Use origin-plugin-starter (https://github.com/fellanH/origin-plugin-starter) for new plugins. This in-tree template will be removed in the next major version.",
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "dependencies": {
-    "@origin/api": "*"
-  }
-}
