--- packages/terminal/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@origin/terminal",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "main": "src/index.tsx",
-  "types": "src/index.tsx",
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite build --watch"
-  },
-  "dependencies": {
-    "@origin/api": "*",
-    "@xterm/xterm": "^5.5.0",
-    "@xterm/addon-fit": "^0.10.0",
-    "@xterm/addon-webgl": "^0.18.0",
-    "@xterm/addon-clipboard": "^0.1.0",
-    "@tauri-apps/api": "^2"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.4.1",
-    "vite": "^7.0.0"
-  }
-}
