--- packages/terminal/src/manifest.json
@@ -1,7 +1,0 @@
-{
-  "id": "com.origin.terminal",
-  "name": "Terminal",
-  "version": "0.0.1",
-  "description": "Full PTY terminal",
-  "icon": ">_"
-}
