--- packages/terminal/src/manifest.ts
@@ -1,11 +1,0 @@
-import type { PluginManifest } from "@origin/api";
-
-export const manifest: PluginManifest = {
-  id: "com.origin.terminal",
-  name: "Terminal",
-  description: "Full PTY terminal",
-  icon: ">_",
-  version: "0.0.1",
-};
-
-export default manifest;
